backend/requirements.txt
alembic==1.18.4
fastapi==0.135.1
# psycopg 3: psycopg2 can raise UnicodeDecodeError when libpq returns non-UTF-8 error text.
psycopg[binary]==3.3.3
pydantic==2.12.5
pydantic-settings==2.13.1
PyJWT[crypto]==2.10.1
python-dotenv==1.2.2
python-multipart==0.0.22
SQLAlchemy==2.0.48
supabase==2.27.0
uvicorn==0.41.0
pytest==8.4.1
httpx==0.28.1
